diminution taille ETX

+++ inc/opendefs.h
@@ -340,7 +340,7 @@ BEGIN_PACK
 typedef struct{
    uint8_t        addr_64b[8];                  // the address of the bottleneck
    uint8_t        counter;                      // the metric use to compute the bottleneck
-   uint16_t       etx;                          // the ETX of the Path to the bottleneck
+   uint8_t        etx;                          // the ETX of the Path to the bottleneck
 } btneck_t;
 END_PACK
